update shadow map resolution, update instanced draw, update attack logic

+++ player.h
@@ -26,7 +26,7 @@
 struct PlayerAction
 {
 	ActionEnum	actionType;
-	UINT		liveTime;
+	float		liveTime;
 };
 
 struct PlayerAttributes
